+++ test/helpers/testHelpers.js
@@ -56,7 +56,9 @@ function checkListeners(child, names, scope) {
 
 pub.getContext = function(processXml, callback) {
   const Context = require('../../lib/Context');
-  transformer.transform(processXml, {}, (err, definitions, moddleContext) => {
+  transformer.transform(processXml, {
+    camunda: require('camunda-bpmn-moddle/resources/camunda')
+  }, (err, definitions, moddleContext) => {
     if (err) return callback(err);
     const context = new Context(contextHelper.getExecutableProcessId(moddleContext), moddleContext, {});
     return callback(null, context);